level2: Set motd attributes correctly on disk images
The copy operation worked, but not the attribute setting,
because the full source path filename would be appended to the
disk image path.

So link all text files into the current directory, and operate
on the relative file names, like was already done for errmsg
and password..

The same error would have happened for sysgo.cfg as well, but it
is currently not being copied anyway.