Update view.asm
Show example of how to drastically increase the speed of the pixel plotter.

Since a process needs to have an 8K free block to plot a single pixel we know that we can just leave that block mapped in until a new one is needed at which time we clear out the old one and map in the new one.

Almost all applications should see a Drastic increase in speed.
See the Pixel routine.