sc6551: Fix hang on overflow error
Fix a bug in sc6551 handling of receive overflows. Before this fix,
the sc6551 driver hangs a level 1 system when data is sent at 19200 bps
with no delays between characters. With the fix, it drops characters and
returns an E$Read error. That's not ideal, but it's better than a hang.